modular reader settings drawer
